requirements.txt
streamlit
tensorflow
numpy
pandas
matplotlib
seaborn
Pillow
scikit-learn
